Add a small header to stage 2, verify it in VBR
A simple header with maigc value, size, and checksum is added to
the stage 2 binary. The installer updates the values and the VBR
verifies that they are sane before jumping into stage 2.

The checksum simply adds up all the integers in the blob and
stores the two's complement, so when doing it again, the result
has to be 0. This is a quite silly algorithm used by some network
protocols, but it has to fit into the VBR.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>