Use BIOS video services instead of accessing video memory
Restructure the TextScreen class to use INT 10 instead of poking
around video memory on it's own. This saves a ton of space on the
one hand, on the other hand, it solves some problems with accessing
out-of-segment memory.

With the now shrunken TextScreen, we can re-use that in MBR and VBR.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>