Replace stage2 header with a blob that is cast to the header struct
The information is filled in by `fatinstall` and the vbr. Previously
the blob was initialized statically, but the compiler decided it can
no longer do that and instead added bulky constructor code that
nobody called. By replacing it with a blob that others fill for us,
we gain a couple dozen bytes of wasted space back.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>